requirements/base.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
midtransclient==1.4.2
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
